+++ source/renderer/zettlr-quicklook.js
@@ -13,7 +13,6 @@
  * END HEADER
  */
 
-const makeTemplate = require('../common/zettlr-template.js')
 const { makeSearchRegEx } = require('../common/zettlr-helpers.js')
 const { trans } = require('../common/lang/i18n.js')
 
@@ -39,10 +38,9 @@ class ZettlrQuicklook {
     * @param {ZettlrBody} parent   Calling object
     * @param {ZettlrFile} file     The file whose content should be displayed
     */
-  constructor (parent, file, standalone = false) {
+  constructor (parent, file) {
     this._parent = parent
     this._file = file
-    this._standalone = standalone
     this._cm = null
     this._window = null
     this._findTimeout = null // Timeout to begin search after
@@ -84,13 +82,7 @@ class ZettlrQuicklook {
     * Load the Quicklook template and prepare everything
     */
   _load () {
-    if (this._standalone) {
-      this._window = $('body')
-    } else {
-      // Process the template and directly call it to only return the HTML
-      let qlcontent = makeTemplate('other', 'quicklook')
-      this._window = $(qlcontent)
-    }
+    this._window = $('body')
 
     this._cm = CodeMirror.fromTextArea(this._window.find('textarea')[0], {
       readOnly: true,
@@ -109,10 +101,6 @@ class ZettlrQuicklook {
     // Apply heading line classes immediately
     this._cm.execCommand('markdownHeaderClasses')
 
-    if (!this._standalone) {
-      this._makeOverlayWindow()
-    }
-
     this._window.find('#searchWhat').attr('placeholder', trans('dialog.find.find_placeholder'))
 
     this._window.find('.close').first().on('click', (e) => {
@@ -127,143 +115,8 @@ class ZettlrQuicklook {
     */
   show () {
     // Standalone windows are pretty easy.
-    if (this._standalone) {
-      $('body').append(this._window)
-      this._cm.refresh()
-      return this
-    }
-    let height = $(window).height()
-    let width = $(window).width()
-    let qlh = height * 0.66 // Two thirds of screen
-    let qlw = width * 0.5
-
-    // Take care of minimum sizes
-    if (qlh < 400) {
-      qlh = 400
-    }
-    if (qlw < 400) {
-      qlw = 400
-    }
-
-    // Somehow the draggable() plugin thinks it's best to set the position
-    // to relative, which then causes the second window to be positioned
-    // NOT where it should but directly beneath the first QL-Window
-    // (respectively its original place before it was moved).
-    this._window.css('position', 'fixed')
-
-    // Set dimensions and positions
-    this._window.css('width', qlw)
-    this._window.css('height', qlh)
-    this._window.css('top', height / 2 - qlh / 2)
-    this._window.css('left', width / 2 - qlw / 2)
-
-    // Append (e.g., show) and set the body to a correct size and give the
-    // CM a first refresh
     $('body').append(this._window)
-    this._window.find('.body').css(
-      'height',
-      (qlh - this._window.find('.title').outerHeight()) + 'px'
-    )
     this._cm.refresh()
-    return this
-  }
-
-  /**
-   * Makes a quicklook window an overlay
-   * @return {void} No return.
-   */
-  _makeOverlayWindow () {
-    // Draggable stuff and resizing is only necessary for not-standalone windows.
-    let toolbarheight = $('#toolbar').outerHeight()
-
-    this._window.draggable({
-      handle: 'div.title',
-      containment: 'document',
-      cursor: '-webkit-grabbing',
-      stack: '.quicklook',
-      drag: (e, ui) => {
-        if (ui.position.top < toolbarheight) {
-          ui.position.top = toolbarheight
-        }
-      },
-      stop: (e, ui) => {
-        this._cm.focus()
-      }
-    })
-
-    this._window.resizable({
-      handles: 'e, se, s, sw, w',
-      containment: 'document',
-      minHeight: 400,
-      minWidth: 400,
-      resize: (e, ui) => {
-        let bar = this._window.find('.title')
-        this._window.find('.body').css('height', (ui.size.height - bar.outerHeight()) + 'px')
-        this._cm.refresh()
-      },
-      stop: (e, ui) => {
-        // Refresh the editor to account for changes in the size.
-        this._cm.refresh()
-        this._cm.focus()
-      }
-    })
-
-    // Bring quicklook window to front on click on the title
-    this._window.find('.title').first().on('click', (e) => {
-      let max
-      let group = $('.quicklook')
-
-      if (group.length < 1) return
-      max = parseInt(group[0].style.zIndex, 10) || 0
-      $(group).each(function (i) {
-        if (parseInt(this.style.zIndex, 10) > max) {
-          max = parseInt(this.style.zIndex, 10)
-        }
-      })
-
-      this._window.css({ 'zIndex': max + 1 })
-    })
-
-    this._window.find('.title').first().on('dblclick', (e) => {
-      this.toggleWindow()
-    })
-
-    // Activate the event listener to pop-out this window
-    this._window.find('.make-standalone').first().on('click', (e) => {
-      global.ipc.send('make-standalone', this._file.hash, (ret) => {
-        // If the new window was successfully opened, close this one.
-        if (ret) this.close()
-      })
-    })
-  }
-
-  /**
-    * Displays visibility of the window's body.
-    * @return {ZettlrQuicklook} Chainability.
-    */
-  toggleWindow () {
-    if (this._window.hasClass('minimize')) {
-      // Restore
-      this._window.removeClass('minimize')
-      this._window.css('height', this._bodyHeight)
-      let bar = this._window.find('.title')
-      this._window.resizable('enable')
-      this._window.find('.body').css(
-        'height',
-        (parseFloat(this._bodyHeight) - bar.outerHeight()) + 'px'
-      )
-      this._window.find('.CodeMirror').css('display', 'block')
-      this._cm.refresh()
-    } else {
-      // Minimize
-      this._window.addClass('minimize')
-      this._bodyHeight = this._window.css('height')
-      this._window.find('.body').css('height', '0px')
-      this._window.resizable('disable')
-      this._window.css('height', '')
-      this._window.find('.CodeMirror').css('display', 'none')
-    }
-
     return this
   }
 
@@ -275,7 +128,6 @@ class ZettlrQuicklook {
     this._window.detach()
     this._cm = null
     this._window = null
-    this._parent.qlsplice(this) // Remove from ql-list in ZettlrBody
   }
 
   // SEARCH FUNCTIONS STOLEN FROM THE ZETTLREDITOR CLASS