+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed the links inside the footnote tooltips. Now they are readable and don't break out of the tooltip anymore.
 - Improvements to the default PDF template.
 - Fixed an error that prevented Zettlr from being able to automatically import language files.
+- Removed the Quicklook overlay windows. Quicklooks now directly become standalone windows.
 
 ## Under the Hood